requirements.txt
fastapi==0.115.8
jinja2==3.1.3
python-multipart==0.0.9
uvicorn==0.27.1
pydantic>=2.10.4
loguru==0.7.2
python-dotenv>=1.0.1
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
langchain-chroma==0.2.2

# Vector store
chromadb==0.6.3

# Browser
playwright==1.50.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-asyncio==0.25.3
httpx==0.28.1 